fix: comprehensive compilation error fixes - Phase 2
Continued systematic resolution of compilation errors across multiple categories:

## Type Mismatch Fixes (E0308 - RESOLVED):
- Fixed UUID conversions (String to Uuid parsing) in 7 API files
- Fixed WithRejection PhantomData issues (2 files)
- Fixed health check return type
- Added flatten() for Option<()> to Option<T> conversions
- Fixed query result type conversions
- **Result**: E0308 errors: 55 → 0 ✅

## Method Not Found Fixes (E0599 - REDUCED):
- Added sqlx::Row trait imports (2 files)
- Fixed setex typo to set_ex in Redis operations
- Fixed iterator and flatten type issues
- **Result**: E0599 errors: 59 → 25

## Service Clone Implementation:
- Added #[derive(Clone)] to 15+ service structs:
  * user_service, athlete_profile_service
  * coaching_recommendation_service, training_plan_service
  * validation_service, recovery_alert_service
  * user_recovery_profile_service, oura_integration_service
  * training_adjustment_service, vision_analysis_service
  * email_notification_service, video services
  * recommendation services
- Added Arc<> wrapping for pose_estimation_service Session field
- Added Clone to PlanStatus enum and OuraApiClient struct

## Additional Fixes:
- Fixed field name conversions (to_vec() for slice to Vec)
- Fixed f32 to f64 casts for confidence fields
- Fixed auth_service_arc wrapper creation in routes
- Fixed REDIS_CACHE_TTL_SECONDS type (usize → u64)

## Progress Summary:
- E0308 (Type mismatches): 55 → 0 ✅
- E0599 (Method not found): 59 → 25 ✅
- E0560 (Struct fields): 28 → 0 ✅
- E0659 (Ambiguous imports): 48 → 0 ✅
- Small error categories: Multiple resolved ✅

Related to Issue #169

+++ ai-coach-api/src/services/daily_recovery_calculation_job.rs
@@ -12,7 +12,7 @@ use crate::services::{Job, RecoveryAlertService, RecoveryAnalysisService};
 
 const BATCH_SIZE: usize = 100;
 const BATCH_DELAY_MS: u64 = 100;
-const REDIS_CACHE_TTL_SECONDS: usize = 3600; // 1 hour
+const REDIS_CACHE_TTL_SECONDS: u64 = 3600; // 1 hour
 const HIGH_FAILURE_RATE_THRESHOLD: f64 = 0.05; // 5%
 
 /// User information for batch processing
